refactor(api): rename src/__tests__ to src/tests
