+++ packages/api/core/src/tests/actions/exercise_equipment.rs
