+++ packages/api/core/src/tests/auth/logout.rs
@@ -1,4 +1,4 @@
-use crate::__tests__::actions::auth as actions;
+use crate::tests::actions::auth as actions;
 use crate::models::User;
 use crate::prelude::*;
